Place depth logging into the one-second loop.
For now, depth logging will take place every one second. This may
later be moved to the GPS update 10Hz function to pair it with
location data when convenient. Logging entries should be sent to
the telemetry log where they can be converted to text files and
further parsed later.